reactive forms

+++ moments/src/app/components/pages/moment/moment.component.ts
@@ -40,10 +40,12 @@ export class MomentComponent implements OnInit {
   }
 
   async removeHandler(id: number) {
-    await this.momentService.removeMoment(id).subscribe();
+    if (id) {
+      await this.momentService.removeMoment(id).subscribe();
 
-    this.messagesService.add(`Momento excluído com sucesso!`);
+      this.messagesService.add(`Momento excluído com sucesso!`);
 
-    this.router.navigate(['/']);
+      this.router.navigate(['/']);
+    }
   }
 }
